# Optimize CreateFeatureOperator