--- a/example.yaml
+++ b/example.yaml
@@ -30,7 +30,7 @@ pipeline:
             infer_datetime_format: true
         - $create_feature:
             feature_name: Date
-            feature_creator: lambda df:df['End Time'].date()
+            feature_creator: lambda df:df['End Time'].dt.date
         - $drop_features:
             feature_names: ['Minutes Awake', 'Number of Awakenings', 'Time in Bed', 'Minutes REM Sleep', 'Minutes Light Sleep', 'Minutes Deep Sleep']
         - $aggregate: