--- a/examples/fitbit_interday/yaml_config/sleep_dataset.yaml
+++ b/examples/fitbit_interday/yaml_config/sleep_dataset.yaml
@@ -14,7 +14,7 @@ pipeline:
         infer_datetime_format: true
     - $create_feature:
         feature_name: Date
-        feature_creator: lambda df:df['End Time'].date()
+        feature_creator: lambda df: df['End Time'].dt.date
     - $aggregate:
         groupby_feature_names: Date
         aggregation_definition: