--- a/examples/withings/yaml_config/sleep_dataset.yaml
+++ b/examples/withings/yaml_config/sleep_dataset.yaml
@@ -14,7 +14,7 @@ pipeline:
         infer_datetime_format: true
     - $create_feature:
         feature_name: Date
-        feature_creator: lambda df:df['to'].date()
+        feature_creator: lambda df:df['to'].dt.date
     - $aggregate:
         groupby_feature_names: Date
         aggregation_definition: