```
def create_new_feature(df):
    return df['Steps'] + df['Steps']
```
```
%%timeit

df['new_feature'] = df.apply(create_new_feature,
                             axis=1,
                             result_type="reduce")

# 1.17 s ± 243 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```
%%timeit

df['new_feature'] = create_new_feature(df)

# 263 µs ± 21.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```